+++ example/lib/main.dart
@@ -1,8 +1,11 @@
-import 'package:flutter/material.dart';
 import 'dart:async';
+import 'dart:collection';
+import 'dart:convert';
 
+import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_netcore_smartech/flutter_netcore_smartech.dart';
+import 'package:flutter_netcore_smartech/model_event.dart';
 
 void main() {
   runApp(MyApp());
@@ -29,16 +32,19 @@ class _MyAppState extends State<MyApp> {
     String resultTracking;
     // Platform messages may fail, so we use a try/catch PlatformException.
     try {
-      platformVersion = await FlutterNetcoreSmartech.platformVersion;
-    } on PlatformException {
-      platformVersion = 'Failed to get platform version.';
-    }
-
-    try{
-      resultTracking = await FlutterNetcoreSmartech.trackingEvent;
+      resultTracking = await FlutterNetcoreSmartech.getNetcoreAppId;
     } on PlatformException {
       resultTracking = 'Failed to tracking event';
     }
+
+    ModelEvent modelEvent = new ModelEvent();
+    modelEvent.eventName = "Add To Cart";
+
+    HashMap<String, Object> payloadData = new HashMap();
+    payloadData["name"] = "Nexus 5";
+    modelEvent.payloadData = jsonEncode(payloadData,
+        toEncodable: FlutterNetcoreSmartech.encoder);
+    FlutterNetcoreSmartech.trackingEvent(modelEvent);
 
     // If the widget was removed from the tree while the asynchronous platform
     // message was in flight, we want to discard the reply rather than calling