+++ CHANGELOG.md
@@ -1,3 +1,27 @@
 ## 0.0.1
 
 * Testing Version
+
+## 0.0.2
+
+* Remove Meta Data in Lib
+
+## 0.0.3
+
+* Add Dynamic Model Event (Name and payload Data)
+
+## 0.0.4
+
+* Add Method get App Id Netcore
+
+## 0.0.5
+
+* Fix Log
+
+## 0.0.6
+
+* Fix Event Tracker
+
+## 0.0.7
+
+* Add Method Tracking in IOS Native Code